refactor(Label.stories): separated the code so that in the story only the following is invoked
